# uucore: num_format: Optimize `format_float_shortest`